CreatesAsset now uses Persistence
I had to change the type of Asset#meta to Object because the underlying
type for the column still needs to be Sequel::Postgres::JSONHash. The
CreatesAsset class is handling that, but with the new attribute stuff if
I set it as a Hash then it gets typecasted away from this type.