Swapping out for HTTP Basic auth rather than Digest
There are a few reasons for this change. I really do like Digest auth in
theory, but I'm having issues getting the client to cache the
authentication response so it doesn't have a 401 request for every
request it wants to make. Rather than dance around this I decided to
simplify things and just use Basic auth.

Tessa is meant to be an infrastructure service so we control the
environment and the clients that it will have. We can force all clients
to use an HTTPS connection which will make HTTP Basic auth just as
secure and less complicated for debugging.